server/requirements.txt
absl-py==2.1.0
aiohttp==3.9.5
aiosignal==1.3.1
anyio==4.4.0
argcomplete==3.5.0
argon2-cffi==23.1.0
argon2-cffi-bindings==21.2.0
arrow==1.3.0
asgiref==3.8.1
asttokens==2.4.1
astunparse==1.6.3
async-lru==2.0.4
async-timeout==4.0.3
attrs==23.2.0
av==12.0.0
Babel==2.15.0
bcrypt==4.2.0
beautifulsoup4==4.12.3
bitarray==2.9.2
bitstring==4.2.3
bleach==6.1.0
Brotli==1.1.0
cachetools==5.3.3
capstone==6.0.0a1
certifi==2024.6.2
cffi==1.16.0
charset-normalizer==3.3.2
chex==0.1.7
colored-traceback==0.4.2
comm==0.2.2
contourpy==1.2.1
cryptography==43.0.1
cycler==0.12.1
datasets==2.20.0
debugpy==1.8.2
decorator==5.1.1
defusedxml==0.7.1
dill==0.3.8
Django==5.1.3
django-filter==24.3
djangorestframework==3.15.2
dm-tree==0.1.8
ecdsa==0.19.0
environs==11.2.1
etils==1.7.0
exceptiongroup==1.2.2
executing==2.0.1
fastjsonschema==2.20.0
filelock==3.15.4
flatbuffers==24.3.25
flax==0.8.5
fonttools==4.53.1
fqdn==1.5.1
frozenlist==1.4.1
fsspec==2024.5.0
gast==0.4.0
google-auth==2.32.0
google-auth-oauthlib==1.0.0
google-pasta==0.2.0
grpcio==1.64.1
h11==0.14.0
h5py==3.11.0
httpcore==1.0.5
httpx==0.27.0
huggingface-hub==0.24.0
idna==3.7
importlib_resources==6.4.0
intelhex==2.3.0
intervaltree==3.1.0
ipykernel==6.29.5
ipython==8.26.0
ipywidgets==8.1.3
isoduration==20.11.0
jax==0.4.30
jaxlib==0.4.30
jedi==0.19.1
Jinja2==3.1.4
joblib==1.4.2
json5==0.9.25
jsonpointer==3.0.0
jsonschema==4.23.0
jsonschema-specifications==2023.12.1
jupyter==1.0.0
jupyter-console==6.6.3
jupyter-events==0.10.0
jupyter-lsp==2.2.5
jupyter_client==8.6.2
jupyter_core==5.7.2
jupyter_server==2.14.2
jupyter_server_terminals==0.5.3
jupyterlab==4.2.3
jupyterlab_pygments==0.3.0
jupyterlab_server==2.27.2
jupyterlab_widgets==3.0.11
keras==2.13.1
kiwisolver==1.4.5
libclang==18.1.1
Mako==1.3.6
Markdown==3.6
markdown-it-py==3.0.0
MarkupSafe==2.1.5
marshmallow==3.23.1
matplotlib==3.9.1
matplotlib-inline==0.1.7
mdurl==0.1.2
mistune==3.0.2
ml-dtypes==0.4.0
msgpack==1.0.8
multidict==6.0.5
multiprocess==0.70.16
mutagen==1.47.0
namex==0.0.8
nbclient==0.10.0
nbconvert==7.16.4
nbformat==5.10.4
nest-asyncio==1.6.0
notebook==7.2.1
notebook_shim==0.2.4
numpy==1.23.4
oauthlib==3.2.2
opencv-python==4.9.0.80
opt-einsum==3.3.0
optax==0.2.1
optree==0.12.1
orbax-checkpoint==0.5.20
overrides==7.7.0
packaging==23.2
pandas==2.2.2
pandocfilters==1.5.1
paramiko==3.5.0
parso==0.8.4
pexpect==4.9.0
pillow==10.3.0
platformdirs==4.2.2
plumbum==1.9.0
prometheus_client==0.20.0
prompt_toolkit==3.0.47
protobuf==4.25.3
psutil==6.0.0
ptyprocess==0.7.0
pure-eval==0.2.2
pwn==1.0
pwntools==4.13.1
pyarrow==17.0.0
pyarrow-hotfix==0.6
pyasn1==0.6.0
pyasn1_modules==0.4.0
pycparser==2.22
pycryptodomex==3.20.0
pyelftools==0.31
Pygments==2.18.0
PyNaCl==1.5.0
pyparsing==3.1.2
pyserial==3.5
PySocks==1.7.1
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
python-json-logger==2.0.7
pytz==2024.1
PyYAML==6.0.1
pyzmq==26.0.3
qtconsole==5.5.2
QtPy==2.4.1
reedsolo==1.7.0
referencing==0.35.1
requests==2.32.3
requests-oauthlib==2.0.0
rfc3339-validator==0.1.4
rfc3986-validator==0.1.1
rich==13.7.1
ROPGadget==7.5
rpds-py==0.19.0
rpyc==6.0.1
rsa==4.9
scikit-learn==1.5.1
scipy==1.13.1
seaborn==0.13.2
Send2Trash==1.8.3
six==1.16.0
sniffio==1.3.1
sortedcontainers==2.4.0
soupsieve==2.5
sqlparse==0.5.2
stack-data==0.6.3
tensorboard==2.13.0
tensorboard-data-server==0.7.2
tensorflow==2.13.1
tensorflow-decision-forests==1.5.0
tensorflow-estimator==2.13.0
tensorflow-hub==0.16.1
tensorflow-io-gcs-filesystem==0.37.1
tensorflowjs==4.20.0
tensorstore==0.1.63
termcolor==2.4.0
terminado==0.18.1
tf-keras==2.15.0
threadpoolctl==3.5.0
tinycss2==1.3.0
tomli==2.0.1
toolz==0.12.1
tornado==6.4.1
tqdm==4.66.4
traitlets==5.14.3
types-python-dateutil==2.9.0.20240316
typing_extensions==4.5.0
tzdata==2024.1
unicorn==2.1.1
unix-ar==0.2.1
uri-template==1.3.0
urllib3==2.2.2
wcwidth==0.2.13
webcolors==24.6.0
webencodings==0.5.1
websocket-client==1.8.0
websockets==12.0
Werkzeug==3.0.3
widgetsnbextension==4.0.11
wrapt==1.16.0
wurlitzer==3.1.1
xxhash==3.4.1
yarl==1.9.4
yt-dlp==2024.4.9
zipp==3.19.2
zstandard==0.23.0
